Add files via upload
This typing speed test is a basic HTML front-end project utilizing HTML, CSS, and JavaScript to evaluate and improve your typing skills. The interface presents a passage of text that you need to type as quickly and accurately as possible. JavaScript dynamically calculates your words per minute (WPM) and accuracy based on your input. The CSS styling ensures a user-friendly and visually appealing experience, while JavaScript manages real-time updates and feedback. The project is ideal for practicing typing speed and accuracy in a straightforward and interactive manner.